CANThrottle_rework
Modified version of the code used during MEBC 2023 by the OCEANOS team. Capped the output rate of CAN packets at 10Hz, so as not to congest the CAN Bus with useless packets. Reworked the included libraries, so as to be easier to understand. 
IMPORTANT: Code tested for usage with Arduino IDE, with TeensyDuino installed, using its included FlexCAN_T4 library for CAN communication.